Uploaded course code; created gitignore; created R project
See folder 'updating_code'

I organized it in a way that we might encourage students to (i.e. having one R project for the entire course with subdirectories for data, plots, and perhaps assignments).

One reason for this is so we can use relative paths for everything and to avoid using the setwd() command which is not considered a good practice.